Improved bundle lock.
Unlock a bundle directly after container creation.
Still keep bundles locked during an image build.
To prevent deletion/GCing of a bundle that is in use by a container check if a container with the same ID exists after locking the bundle.
This way a detached container's bundle dependency would be respected as well.

Signed-off-by: Max Goltzsche <max.goltzsche@gmail.com>